requirements.txt
coverage==4.5.2
codecov==2.0.15
nose==1.3.7
mock==2.0.0
GitPython==2.1.11
PyYAML<=3.13,>=3.12
ansible==2.7.5
future==0.17.1
twine==1.12.1
click==7.0